requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Charm-Crypto==0.50
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
hypothesis==6.14.0
idna==2.10
importlib-metadata==4.5.0
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.0
python-dotenv==0.18.0
requests==2.25.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.19
starlette==0.14.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
zipp==3.4.1
